Allow browsers to use distribution files directly
Remove dependencies on lodash and resize-observer-polyfill. We only used
throttle() from lodash, and in a very narrow use case with an obvious
implementation. ResizeObserver is now standard enough that we can live
without the polyfill.

Append .js extensions to imports, now that TypeScript allows that
slightly quirky feature, so that browsers can resolve them.